General overview of entity class

+++ src/main/java/se/iths/projektarbetekomplexjava/controller/BookController.java
@@ -4,8 +4,10 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
+import se.iths.projektarbetekomplexjava.entity.Author;
 import se.iths.projektarbetekomplexjava.entity.Book;
 import se.iths.projektarbetekomplexjava.entity.Customer;
+import se.iths.projektarbetekomplexjava.exception.NotFoundException;
 import se.iths.projektarbetekomplexjava.service.BookService;
 
 import java.util.List;
@@ -22,44 +24,49 @@ public class BookController {
     }
 
     @PostMapping("/addbook")
-    //@PreAuthorize("hasRole('ADMIN')")
+    @PreAuthorize("hasRole('ADMIN')")
     public ResponseEntity<Book> addBook(@RequestBody Book book) {
         Book addedBook = bookService.addBook(book);
         return new ResponseEntity<>(addedBook, HttpStatus.CREATED);
     }
 
-    @GetMapping("/getbookbyisbn")
+    @GetMapping("/getbookbyisbn/{ISBN13}")
     public ResponseEntity<Book> getByISBN(@PathVariable String ISBN13) {
         Book searchBook = bookService.getBookByISBN13(ISBN13);
         return new ResponseEntity<>(searchBook, HttpStatus.OK);
     }
 
-    @GetMapping("/getbookbyauthor")
-    public ResponseEntity<List<Book>> getBooksByAuthor(@PathVariable String firstName, String lastName) {
-        List<Book> booksByAuthor = bookService.getBooksByAuthor(firstName, lastName);
-        return new ResponseEntity<>(booksByAuthor, HttpStatus.OK);
-    }
+//    @GetMapping("/getbookbyauthor")
+//    public ResponseEntity<List<Book>> getBooksByAuthor(@PathVariable String firstName, String lastName) {
+//        List<Book> booksByAuthor = bookService.getBooksByAuthor(firstName, lastName);
+//        return new ResponseEntity<>(booksByAuthor, HttpStatus.OK);
+//    }
 
-    @GetMapping("/getbookbytitle")
+    @GetMapping("/getbookbytitle/{title}")
     public ResponseEntity<List<Book>> getBookByTitle(@PathVariable String title) {
         List<Book> booksByTitle = bookService.getBookByTitle(title);
         return new ResponseEntity<>(booksByTitle, HttpStatus.OK);
     }
 
-    @GetMapping("/getbookbylanguage")
+    @GetMapping("/getbookbylanguage/{language}")
     public ResponseEntity<List<Book>> getBookByLanguage(@PathVariable String language) {
         List<Book> booksByLanguage = bookService.getBooksByLanguage(language);
         return new ResponseEntity<>(booksByLanguage, HttpStatus.OK);
     }
 
-    @GetMapping("/getbookbyid")
+    @GetMapping("/getbookbyid/{id}")
     public ResponseEntity<Optional<Book>> getBookById(@PathVariable Long id) {
         Optional<Book> bookById = bookService.findByBookId(id);
         return new ResponseEntity<>(bookById, HttpStatus.OK);
     }
 
     @DeleteMapping("/deletebook/{id}")
+    @PreAuthorize("hasRole('ADMIN')")
     public ResponseEntity<Void> removeBoook(@PathVariable Long id) {
+        Optional<Book> foundBook = bookService.findByBookId(id);
+        if (foundBook.isEmpty()){
+            throw new NotFoundException("No data available of book ID: " + id);
+        }
         bookService.removeBook(id);
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
@@ -71,12 +78,14 @@ public class BookController {
     }
 
     @PutMapping("/updatebook")
+    @PreAuthorize("hasRole('ADMIN')")
     public ResponseEntity<Book> updateBook(@RequestBody Book book) {
         Book updateBook = bookService.updateBook(book);
         return new ResponseEntity<>(updateBook, HttpStatus.OK);
     }
 
     @PatchMapping("/updatebookinformation")
+    @PreAuthorize("hasRole('ADMIN')")
     public ResponseEntity<Book> updateBook1(@RequestBody Book book) {
         Book updateBook = bookService.updateBook1(book);
         return new ResponseEntity<>(updateBook, HttpStatus.OK);
